langgraph-app/requirements.txt
langchain>=0.3.0
langgraph>=0.3.0
langchain-community>=0.3.0
langchain-google-vertexai>=2.0.0
pydantic>=2.7.1
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.108.0
uvicorn>=0.25.0
redis>=5.0.0
pillow>=10.0.0
google-cloud-aiplatform>=1.44.0 